Add methods to remove directly from the node
It's not longer necessary to use the parent node to remove a node.
We can request removing the node from itself using remove_me or sibling using remove_previous_sibling / remove_next_sibling.
The parent node must be defined, but the definition is done when node are inserted of from parsed XML, so, it almost standard.